end the war between uncrustify and cpplint
